requirements.txt
aiohttp==2.2.5
async-timeout==2.0.0
cffi==1.11.5
chardet==3.0.4
discord.py==1.0.0a0
idna==2.6
multidict==4.1.0
pycparser==2.18
PyMySQL==0.8.0
PyNaCl==1.1.2
six==1.11.0
SQLAlchemy==1.2.5
websockets==4.0.1
yarl==1.1.1
